nanopi-r5c: apply static MAC addresses for both Ethernet adapters
U-Boot generates and assigns two MAC addresses for Rockchip SBCs with ethernet0 and ethernet1 aliases pointing to repsective device nodes.

The PCIe child nodes for both 2.5 Gbit Ethernet adapters with the needed "local-mac-address" properties are added with this patch, as well as the respective aliases.

This commit applies as well default Ethernet LED triggers.

Signed-off-by: MichaIng <micha@dietpi.com>