nanopi-r4s: apply static MAC addresses for both Ethernet adapters
U-Boot generates and assigns two MAC addresses for Rockchip SBCs with ethernet0 and ethernet1 aliases pointing to repsective device nodes.

The PCIe child node for the PCIe Ethernet adapter with the needed "local-mac-address" property is added with this patch, as well as the respective alias.

This commit additionally applies default Ethernet LED triggers.

Signed-off-by: MichaIng <micha@dietpi.com>