mainline-kernel: align enabled WiFi drivers
- Enable all RTW88 mainline drivers
- Disable all related 3rd party drivers
- Merge and sort all related config keys
- Align for all mainline Rockchip/Allwinner/Amlogic configs

As an exception, keep using the out-of-tree driver for RTL8723DS, which performs much better on Linux 6.12. On Linux 6.16, the mainline driver runs as smooth, but throws a warning invalid efuse MAC, assigning a random one instead.  On ROCK Pi S, the out-of-tree driver does not assign a static MAC etiher, but only the last 3 blocks are random, the first 3 remain static. On the Gateway GZ80X, as heard from someone else, it seems to apply a completely static MAC. In any case, it can obviously better obtain MAC address info hardware, respectively do so in more cases.

Signed-off-by: MichaIng <micha@dietpi.com>